Move download button to own component
